% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_days_best.R
\name{get_days_best}
\alias{get_days_best}
\title{Returns a Text-Formatted List of a Day's Best Lines}
\usage{
get_days_best(
  game_date = character(0),
  sport = character(0),
  market = "Moneyline",
  league = character(0)
)
}
\arguments{
\item{game_date}{The date you want to filter games by - format as "YYYY-MM-DD"}

\item{sport}{The sport you want lines for.}

\item{market}{The market you want lines for. Defaults to Moneyline. See ?get_markets for instructions on how to get a list of markets.}

\item{startDateAfter}{Filter by games that start after a date. Format dates as "YYYY-MM-DD" for example "2021-10-31"}
}
\value{
Returns a dataframe with the requested markets.
}
\description{
To use this function you will need an API key and you must pass your API key to the 'set_OddsJam_api_key()' function.
}
\details{
Type ?set_OddsJam_api_key into the console for more info.

'key' is the only parameter that is required.
}
