% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gameIds.R
\name{get_gameIds}
\alias{get_gameIds}
\title{List gameIds}
\usage{
get_gameIds(
  oj_url = "https://api-external.oddsjam.com/api/feed/games/",
  key = .OddsJamEnv$data$apikey,
  page = character(0),
  sport = character(0),
  league = character(0),
  isLive = FALSE,
  startDateBefore = character(0),
  startDateAfter = character(0)
)
}
\arguments{
\item{key}{Your api key is automatically put here if you have passed it to the set_OddsJam_api_key function.}

\item{page}{The page of the response you want. Note: the API is paginated and only sends 500 odds at a time.}

\item{sport}{The sport you want. Defaults to baseball. For a list of available sports see: https://developer.oddsjam.com/leagues}

\item{league}{Filter by league. You can get a list of available leagues with the get_leagues() function.}

\item{isLive}{Filters odds for live games. Defaults to FALSE. Using TRUE will override startDateBefore and startDateAfter.}

\item{startDateBefore}{Filter by games that start before a date. Format dates as "YYYY-MM-DD" for example "2021-10-31".}

\item{startDateAfter}{Filter by games that start after a date. Format dates as "YYYY-MM-DD" for example "2021-10-31"}
}
\value{
Returns a dataframe with the requested markets.
}
\description{
To use this function you will need an API key and you must pass your API key to the 'set_OddsJam_api_key()' function.
}
\details{
Type ?set_OddsJam_api_key into the console for more info.

'key' is the only parameter that is required.
}
\examples{
get_gameIds()
}
