% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leagues.R
\name{get_leagues}
\alias{get_leagues}
\title{List Available Leagues for a Sport}
\usage{
get_leagues(
  oj_url = "https://api-external.oddsjam.com/api/feed/leagues/",
  key = .OddsJamEnv$data$apikey,
  sport = "baseball",
  isLive = FALSE
)
}
\arguments{
\item{key}{Your api key is automatically put here if you have passed it to the set_OddsJam_api_key function.}

\item{sport}{The sport you want. Defaults to baseball. For a list of available sports see: https://developer.oddsjam.com/leagues}

\item{isLive}{Filters odds for live games. Defaults to FALSE.}
}
\value{
Returns a dataframe with the requested leagues.
}
\description{
To use this function you will need an API key and you must pass your API key to the 'set_OddsJam_api_key()' function.
}
\details{
Type ?set_OddsJam_api_key into the console for more info.

'key' is the only parameter that is required.
}
\examples{
get_leagues()
}
