% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_markets.R
\name{get_markets}
\alias{get_markets}
\title{List Available Markets}
\usage{
get_markets(
  oj_url = "https://api-external.oddsjam.com/api/feed/markets/",
  key = .OddsJamEnv$data$apikey,
  page = character(0),
  gameId = character(0),
  isLive = TRUE,
  time_zone = "America/New_York"
)
}
\arguments{
\item{key}{Your api key is automatically put here if you have passed it to the set_OddsJam_api_key function.}

\item{page}{The page of the response you want. Note: the API is paginated and only sends 500 odds at a time.}

\item{gameId}{The OddsJam game ID you want to receive odds for (e.g. 37621).}

\item{isLive}{Filters odds for live games. Defaults to TRUE.}

\item{time_zone}{You can get the times in your desired time zone. Defaults to Eastern Time. To see a list of useable time zones type OlsonNames() into the console.}
}
\value{
Returns a dataframe with the requested markets.
}
\description{
To use this function you will need an API key and you must pass your API key to the 'set_OddsJam_api_key()' function.
}
\details{
Type ?set_OddsJam_api_key into the console for more info.

'key' is the only parameter that is required.
}
\examples{
get_markets()
}
