% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_OddsJam_api_key.R
\name{set_OddsJam_api_key}
\alias{set_OddsJam_api_key}
\title{Pass your secret API key to this function}
\usage{
set_OddsJam_api_key(key)
}
\arguments{
\item{key}{Your secret API key. Do not share this with anyone.}
}
\description{
Pass your secret API key to this function
}
\examples{
set_OddsJam_api_key("1a1a1a11-2222-33b3-ccc4-55d5dd5555d5")

}
